Move indexer to its own folder
