+++ internal/repository/edgedevice/edgedevice.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/project-flotta/flotta-operator/api/v1alpha1"
-	indexer "github.com/project-flotta/flotta-operator/internal/repository"
+	"github.com/project-flotta/flotta-operator/internal/indexer"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
