+++ main.go
@@ -20,20 +20,19 @@ import (
 	"context"
 	"crypto/x509"
 	"fmt"
-	"github.com/project-flotta/flotta-operator/internal/informers"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"strings"
 
-	"github.com/project-flotta/flotta-operator/internal/configmaps"
-	"github.com/project-flotta/flotta-operator/internal/devicemetrics"
-	indexer "github.com/project-flotta/flotta-operator/internal/repository"
-
 	"github.com/kelseyhightower/envconfig"
 	routev1 "github.com/openshift/api/route/v1"
+	"github.com/project-flotta/flotta-operator/internal/configmaps"
+	"github.com/project-flotta/flotta-operator/internal/devicemetrics"
 	"github.com/project-flotta/flotta-operator/internal/images"
+	"github.com/project-flotta/flotta-operator/internal/indexer"
+	"github.com/project-flotta/flotta-operator/internal/informers"
 	"github.com/project-flotta/flotta-operator/internal/k8sclient"
 	"github.com/project-flotta/flotta-operator/internal/metrics"
 	"github.com/project-flotta/flotta-operator/internal/repository/edgedeployment"
